+++ Studio.Api.Model/Users/User.cs
@@ -9,19 +9,19 @@ namespace Studio.Api.Model.Users
         public string Email { get; set; }
         public string Name { get; set; }
         public string RestrictedStatus { get; set; }
-        public int Id { get; set; }
+        public long Id { get; set; }
     }
 
     public class ProjectUsersList
     { 
         public List<User> ProjectUsers { get; set; }
-        public int TotalCount { get; set; }
+        public long TotalCount { get; set; }
     }
 
     public class SessionUsersList
     {
         public List<User> SessionUsers { get; set; }
-        public int TotalCount { get; set; }
+        public long TotalCount { get; set; }
     }
 
     public class UpdateUserRestrictedStatusRequest
